Restore modules and move httputil into esplora
