+++ go.mod
@@ -2,25 +2,25 @@ module github.com/tdex-network/tdex-daemon
 
 go 1.14
 
+replace github.com/tdex-network/tdex-daemon/pkg/explorer => ./pkg/explorer
+
 require (
 	github.com/btcsuite/btcd v0.21.0-beta
 	github.com/btcsuite/btcutil v1.0.2
-	github.com/desertbit/timer v0.0.0-20180107155436-c41aec40b27f // indirect
 	github.com/dgraph-io/badger/v2 v2.2007.2
 	github.com/gogo/protobuf v1.2.1
 	github.com/google/uuid v1.1.2
 	github.com/grpc-ecosystem/go-grpc-middleware v1.0.0
 	github.com/improbable-eng/grpc-web v0.13.0
-	github.com/magiconair/properties v1.8.1
 	github.com/prometheus/client_golang v0.9.3
-	github.com/rs/cors v1.7.0 // indirect
 	github.com/shopspring/decimal v1.2.0
 	github.com/sirupsen/logrus v1.7.0
 	github.com/soheilhy/cmux v0.1.4
 	github.com/sony/gobreaker v0.4.1
 	github.com/spf13/viper v1.7.1
 	github.com/stretchr/testify v1.7.0
-	github.com/tdex-network/tdex-protobuf v0.0.0-20210322191920-4609ea418f3a
+	github.com/tdex-network/tdex-daemon/pkg/explorer v0.0.0-20210317222720-0adcc0c066a5
+	github.com/tdex-network/tdex-protobuf v0.0.0-20210324125236-78f6c6c0618b
 	github.com/thanhpk/randstr v1.0.4
 	github.com/timshannon/badgerhold/v2 v2.0.0-20201016201833-94bc303c76d4
 	github.com/tyler-smith/go-bip39 v1.1.0 // indirect
