+++ pkg/explorer/go.mod
@@ -1,0 +1,11 @@
+module github.com/tdex-network/tdex-daemon/pkg/explorer
+
+go 1.15
+
+require (
+	github.com/btcsuite/btcd v0.21.0-beta
+	github.com/magiconair/properties v1.8.4
+	github.com/stretchr/testify v1.7.0
+	github.com/tdex-network/tdex-daemon v0.2.1-0.20210317215952-93a47434cc41
+	github.com/vulpemventures/go-elements v0.1.1
+)
